updated apiserviec and deleted merge left overs
slight fix to dashboard
