+++ components/App/ArbSwap/Dashboard/Offers/OffersSent.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Offer } from '@/constants/Offer'
 
-function OffersSent({ offers }: { offers: Offer[] }) {
+function OffersSent({ offers }: Readonly<{ offers: Offer[] }>) {
   return (
     <div className="text-black">
       {offers.map((offer) => {
